Time: 7 ms (24.34%), Space: 10.4 MB (78.5%) - LeetHub

+++ 2800-minimum-string-length-after-removing-substrings/2800-minimum-string-length-after-removing-substrings.cpp
@@ -3,7 +3,7 @@ public:
     int minLength(string s) {
         if(s.size() <= 1) return s.size();
         string stack = {s[0]};
-        for(int i = 1;i < s.size();i++){
+        for(short i = 1;i < s.size();i++){
             if(!stack.empty() && ((stack.back() == 'A' && s[i] == 'B') || (stack.back() == 'C' && s[i] == 'D'))) stack.pop_back();
             else stack.push_back(s[i]);
         }
